+++ CHANGELOG.rst
@@ -34,6 +34,7 @@ New Features
   (:issue:`575`, :pr:`675`, :pr:`678`) `Aaron Spring`_.
 - Allow ``lead`` as ``float`` also if ``calendar="360_day"`` or ``lead.attrs["units"]``
   not in ``["years","seasons","months"]``. (:issue:`564`, :pr:`675`) `Aaron Spring`_.
+- Implement :py:func:`~climpred.metrics._less` (:issue:`239`, :pr:`687`) `Aaron Spring`_.
 
 Documentation
 -------------